feat(tool): migrate KeystoreFactory SolidityNode and DBConvert into FullNode
