+++ common/src/main/java/org/tron/common/parameter/CommonParameter.java
@@ -199,7 +199,20 @@ public class CommonParameter {
   //If you are running a solidity node for java tron, this flag is set to true
   @Getter
   @Setter
+  @Parameter(names = {"--solidity"}, description = "running a solidity node for java tron")
   public boolean solidityNode = false;
+  //If you are running KeystoreFactory, this flag is set to true
+  @Getter
+  @Setter
+  @Parameter(names = {"--keystore"}, description = "running KeystoreFactory")
+  public boolean keystore = false;
+
+  //If you are running leveldb to rocksdb convert tool, this flag is set to true
+  @Getter
+  @Setter
+  @Parameter(names = {"--db-convert"}, description = "running leveldb to rocksdb convert")
+  public boolean dbConvert = false;
+
   @Getter
   @Setter
   public int rpcPort;
