+++ framework/build.gradle
@@ -153,7 +153,9 @@ def binaryRelease(taskName, jarName, mainClass) {
         from(sourceSets.main.output) {
             include "/**"
         }
-
+        // explicit_dependency
+        dependsOn (project(':actuator').jar, project(':consensus').jar, project(':chainbase').jar,
+                project(':crypto').jar, project(':common').jar, project(':protocol').jar)
         from {
             configurations.runtimeClasspath.collect {
                 it.isDirectory() ? it : zipTree(it)
@@ -218,36 +220,12 @@ startScripts.enabled = false
 run.enabled = false
 tasks.distTar.enabled = false
 
-createScript(project, 'org.tron.program.SolidityNode', 'SolidityNode')
 createScript(project, 'org.tron.program.FullNode', 'FullNode')
-createScript(project, 'org.tron.program.KeystoreFactory', 'KeystoreFactory')
-createScript(project, 'org.tron.program.DBConvert', 'DBConvert')
 
-def releaseBinary = hasProperty('binaryRelease') ? getProperty('binaryRelease') : 'true'
-def skipSolidity = hasProperty('skipSolidity') ? true : false
-def skipKeystore = hasProperty('skipKeystore') ? true : false
-def skipConvert = hasProperty('skipConvert') ? true : false
-def skipAll = hasProperty('skipAll') ? true : false
-if (releaseBinary == 'true') {
-    artifacts {
-        archives(binaryRelease('buildFullNodeJar', 'FullNode', 'org.tron.program.FullNode'))
-    }
-    if (!skipAll) {
-        if (!skipSolidity) {
-            artifacts {
-                archives(binaryRelease('buildSolidityNodeJar', 'SolidityNode', 'org.tron.program.SolidityNode'))}
-        }
-        if (!skipKeystore) {
-            artifacts {
-                archives(binaryRelease('buildKeystoreFactoryJar', 'KeystoreFactory', 'org.tron.program.KeystoreFactory'))}
-        }
-        if (!skipConvert) {
-            artifacts {
-                archives(binaryRelease('buildDBConvertJar', 'DBConvert', 'org.tron.program.DBConvert'))}
-        }
-    }
-
+artifacts {
+    archives(binaryRelease('buildFullNodeJar', 'FullNode', 'org.tron.program.FullNode'))
 }
+
 
 task copyToParent(type: Copy) {
     into "../build/distributions"
